+++ www/SQLitePlugin.js
@@ -315,7 +315,7 @@
   };
 
   SQLitePluginTransaction.prototype.start = function() {
-    var err, error1;
+    var err;
     try {
       this.fn(this);
       this.run();
@@ -402,7 +402,7 @@
     tx = this;
     handlerFor = function(index, didSucceed) {
       return function(response) {
-        var err, error1;
+        var err;
         try {
           if (didSucceed) {
             tx.handleStatementSuccess(batchExecutes[index].success, response);