+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Changes
 
-###### cordova-sqlite-legacy-express-core 1.0.0-pre1
+###### cordova-sqlite-legacy-express-core 1.0.0-pre2
 
+- selfTest simulate scenario in BUG litehelpers/Cordova-sqlite-storage#666 (also includes string test and test of effects of location reload/change in this version branch, along with another internal check)
 - Drop engine constraints in package.json & plugin.xml (in this version branch)
 - Support macOS platform with builtin libsqlite3.dylib framework in this version branch