+++ CHANGES.md
@@ -1,8 +1,9 @@
 # Changes
 
-###### cordova-sqlite-legacy-express-core 1.0.0-pre2
+###### cordova-sqlite-legacy-express-core 1.0.0-pre3
 
-- selfTest simulate scenario in BUG litehelpers/Cordova-sqlite-storage#666 (also includes string test and test of effects of location reload/change in this version branch, along with another internal check)
+- Workaround solution to BUG litehelpers/Cordova-sqlite-storage#666 (hanging transaction in case of location reload/change)
+- selfTest simulate scenario & test solution to BUG litehelpers/Cordova-sqlite-storage#666 (also includes string test and test of effects of location reload/change in this version branch, along with another internal check)
 - Drop engine constraints in package.json & plugin.xml (in this version branch)
 - Support macOS platform with builtin libsqlite3.dylib framework in this version branch