+++ CHANGES.md
@@ -1,10 +1,11 @@
 # Changes
 
-###### cordova-sqlite-legacy-express-core 1.0.0-pre3
+###### cordova-sqlite-legacy-express-core 1.0.0-pre4
 
 - Workaround solution to BUG litehelpers/Cordova-sqlite-storage#666 (hanging transaction in case of location reload/change)
 - selfTest simulate scenario & test solution to BUG litehelpers/Cordova-sqlite-storage#666 (also includes string test and test of effects of location reload/change in this version branch, along with another internal check)
 - Drop engine constraints in package.json & plugin.xml (in this version branch)
+- Remove Lawnchair adapter from this version branch
 - Support macOS platform with builtin libsqlite3.dylib framework in this version branch
 
 ## 1.2.2