+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## cordova-sqlite-storage 2.1.0-browser-wip1
+
+TBD
+
 ## cordova-sqlite-storage 2.0.3
 
 - Drop engines rule from package.json